agent-python/requirements.txt
requests
mss
Pillow
pystray
pyinstaller
